web: organise components into folders
