+++ ui/src/App.js
@@ -5,7 +5,7 @@ import Alert from '@mui/material/Alert';
 import {LicenseManager} from 'ag-grid-enterprise';
 import React, {useState, useMemo} from 'react';
 import {BrowserRouter as Router, Navigate, Route, Routes} from 'react-router-dom';
-import {AuthContext} from './auth-context';
+import {AuthContext} from './contexts/auth-context';
 import LoginForm from './components/auth/LoginForm';
 import AppContent from './components/containers/AppContent';
 import DiverAdd from './components/data-entities/diver/DiverAdd';
@@ -13,13 +13,13 @@ import DiverList from './components/data-entities/diver/DiverList';
 import LocationAdd from './components/data-entities/location/LocationAdd';
 import LocationList from './components/data-entities/location/LocationList';
 import LocationView from './components/data-entities/location/LocationView';
-import ObservableItemAdd from './components/data-entities/ObservableItemAdd';
-import ObservableItemEdit from './components/data-entities/ObservableItemEdit';
-import ObservableItemList from './components/data-entities/ObservableItemList';
-import ObservableItemView from './components/data-entities/ObservableItemView';
-import SiteEdit from './components/data-entities/SiteEdit';
-import SiteList from './components/data-entities/SiteList';
-import SiteView from './components/data-entities/SiteView';
+import ObservableItemAdd from './components/data-entities/observable-item/ObservableItemAdd';
+import ObservableItemEdit from './components/data-entities/observable-item/ObservableItemEdit';
+import ObservableItemList from './components/data-entities/observable-item/ObservableItemList';
+import ObservableItemView from './components/data-entities/observable-item/ObservableItemView';
+import SiteEdit from './components/data-entities/site/SiteEdit';
+import SiteList from './components/data-entities/site/SiteList';
+import SiteView from './components/data-entities/site/SiteView';
 import SurveyEdit from './components/data-entities/survey/SurveyEdit';
 import SurveyList from './components/data-entities/survey/SurveyList';
 import SurveyView from './components/data-entities/survey/SurveyView';
