+++ ui/src/components/data-entities/observable-item/ObservableItemList.js
@@ -3,8 +3,8 @@
 import React, {useEffect, useState} from 'react';
 import {Box, Button, Typography} from '@mui/material';
 import {Navigate, NavLink} from 'react-router-dom';
-import {getResult} from '../../axios/api';
-import LoadingOverlay from '../overlays/LoadingOverlay';
+import {getResult} from '../../../api/api';
+import LoadingOverlay from '../../overlays/LoadingOverlay';
 import {AgGridColumn, AgGridReact} from 'ag-grid-react';
 import {Add} from '@mui/icons-material';
