+++ ui/src/components/data-entities/survey/SurveyView.js
@@ -3,7 +3,7 @@ import {Edit} from '@mui/icons-material';
 import React, {useEffect, useState} from 'react';
 import {useParams} from 'react-router';
 import {NavLink} from 'react-router-dom';
-import {getEntity} from '../../../axios/api';
+import {getEntity} from '../../../api/api';
 import EntityContainer from '../../containers/EntityContainer';
 import CustomTextInput from '../../input/CustomTextInput';
