+++ ui/src/components/import/DataSheetView.js
@@ -12,8 +12,8 @@ import 'ag-grid-community/dist/styles/ag-grid.css';
 import {AgGridColumn, AgGridReact} from 'ag-grid-react';
 import {PropTypes} from 'prop-types';
 import {NavLink} from 'react-router-dom';
-import {getDataJob, submitIngest, updateRows, validateJob} from '../../axios/api';
-import {extendedMeasurements, measurements} from '../../constants';
+import {getDataJob, submitIngest, updateRows, validateJob} from '../../api/api';
+import {extendedMeasurements, measurements} from '../../common/constants';
 import LoadingOverlay from '../overlays/LoadingOverlay';
 import AlertDialog from '../ui/AlertDialog';
 import FindReplacePanel from './panel/FindReplacePanel';
