+++ ui/src/components/job/JobView.js
@@ -13,7 +13,7 @@ import FileDownload from 'js-file-download';
 import React, {useEffect, useState} from 'react';
 import {useParams} from 'react-router';
 import {Link} from 'react-router-dom';
-import {getEntity, originalJobFile} from '../../axios/api';
+import {getEntity, originalJobFile} from '../../api/api';
 import EntityContainer from '../containers/EntityContainer';
 
 const JobView = () => {
